% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalize-package.R
\docType{package}
\name{anomalize_package}
\alias{anomalize_package}
\alias{anomalize_package-package}
\title{anomalize: Tidy anomaly detection}
\description{
anomalize: Tidy anomaly detection
}
\details{
The \code{anomalize} package enables a tidy workflow for detecting anomalies in data.
The main functions are \code{time_decompose()}, \code{anomalize()}, and \code{time_recompose()}.
When combined, it's quite simple to decompose time series, detect anomalies,
and create bands separating the "normal" data from the anomalous data.

To learn more about \code{anomalize}, start with the vignettes:
\code{browseVignettes(package = "anomalize")}
}
