% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse_cran_downloads.R
\docType{data}
\name{tidyverse_cran_downloads}
\alias{tidyverse_cran_downloads}
\title{Downloads of various "tidyverse" packages from CRAN}
\format{
A \code{grouped_tbl_time} object with 6,375 rows and 3 variables:
\describe{
\item{date}{Date of the daily observation}
\item{count}{Number of downloads that day}
\item{package}{The package corresponding to the daily download number}
}
}
\source{
The package downloads come from CRAN by way of the \code{cranlogs} package.
}
\usage{
tidyverse_cran_downloads
}
\description{
A dataset containing the daily download counts from 2017-01-01 to 2018-03-01
for the following tidyverse packages:
\itemize{
\item \code{tidyr}
\item \code{lubridate}
\item \code{dplyr}
\item \code{broom}
\item \code{tidyquant}
\item \code{tidytext}
\item \code{ggplot2}
\item \code{purrr}
\item \code{stringr}
\item \code{forcats}
\item \code{knitr}
\item \code{readr}
\item \code{tibble}
\item \code{tidyverse}
}
}
\keyword{datasets}
