% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_template.R
\name{set_time_scale_template}
\alias{set_time_scale_template}
\alias{get_time_scale_template}
\alias{time_scale_template}
\title{Get and modify time scale template}
\usage{
set_time_scale_template(data)

get_time_scale_template()

time_scale_template()
}
\arguments{
\item{data}{A \code{tibble} with a "time_scale", "frequency", and "trend" columns.}
}
\description{
Get and modify time scale template
}
\details{
Used to get and set the time scale template, which is used by \code{time_frequency()}
and \code{time_trend()} when \code{period = "auto"}.
}
\examples{

get_time_scale_template()

set_time_scale_template(time_scale_template())

}
\seealso{
\code{\link[=time_frequency]{time_frequency()}}, \code{\link[=time_trend]{time_trend()}}
}
